Start cleaning up the way xyz are changed
